move logging initialization to hoddb.go
